AMP-8734
git-svn-id: file:///svn/branches/AMP_1_14_RELEASE/amp@12827 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ WEB-INF/src/org/digijava/module/widget/action/ShowSectorByDonorChart.java
@@ -121,7 +121,7 @@ public class ShowSectorByDonorChart extends Action {
         }
         //HEIGHT
         if (form.getImageHeight() == null) {
-            opt.setHeight(new Integer(400));
+            opt.setHeight(new Integer(480));
         } else {
             opt.setHeight(form.getImageHeight());
         }
